// src/types/infrastructure/database/database.ts
export interface PsocMajorGroup {
  code: string;
  title: string;
  created_at: string;
}

export interface PsocSubMajorGroup {
  code: string;
  title: string;
  major_code: string;
  created_at: string;
}

export interface PsocMinorGroup {
  code: string;
  title: string;
  sub_major_code: string;
  created_at: string;
}

export interface PsocUnitGroup {
  code: string;
  title: string;
  minor_code: string;
  created_at: string;
}

export interface PsocUnitSubGroup {
  code: string;
  title: string;
  unit_code: string;
  created_at: string;
}